# doc/api.yaml
openapi: 3.0.0
servers:
  - url: "http://localhost:3000"
info:
  title: WASAText
  description: WASAText, the skeleton project for Web and Software architecture course
  version: 1.0.0
...
paths:
...
  /login:
    post:
      security:
        - bearerAuth: []
      tags:
        - Login
      summary: logs in the  user
      description: |-
        if the user does not exist it will be created
        else it will be logged in; either way the userId is returned
      operationId: DOLOGIN
      requestBody:
        description: Username to be used for loggin in
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Name"

        required: true

      responses:
        "201":
          description: user log in action succesful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

        "400":
          $ref: "#/components/responses/BadRequest"

        "403":
          $ref: "#/components/responses/Forbidden"

        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{userId}:

    get:
      security:
      - bearerAuth: []
      tags:
        -  User
      summary: gets the user profile
      description: |-
        retrieves the elements
        of the user profile
      parameters:
      - $ref: "#/components/parameters/userId"
      operationId: GETUSER
      responses:
        "200":
          description: conversation retrieved succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"


    put:
      security:
        - bearerAuth: []
      tags:
        - User
      summary: sets the name of the user
      description: |-
        sets the name that
        the user will be seen with
      parameters:

        - $ref: "#/components/parameters/userId"

      operationId: SETMYUSERNAME
      requestBody:
        description: creates/changes your own username
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Name"
      responses:
        "204":
          description: username updated succesfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  parameters:
    userId:
      description: identifier for a single user
      schema:
        $ref: "#/components/schemas/Identifier"
      name: userId
      in: path
      required: true
...
  responses:
      BadRequest:
        description: syntax of the request was not compliant with docs, check parameters
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Error'
      Forbidden:
        description: server will not accept request due to lacking of client permissions
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Error'
...
      InternalServerError:
        description: the server encountered an error while processing the request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    Photo:
      type: object
      required: [photoUrl]
      properties:
        photoUrl:
          type: string
          format: url
          minLength: 10
          maxLength: 2048
          pattern: '^https?://.*$'
          description: New photo URL for the user
    Name:
      type: object
      properties:
        name:
          type: string
          example: filippohd
          pattern: "^[a-z0-9]*?$"
          minLength: 1
          maxLength: 12
      required:
        - name
    Identifier:
      title: identifier
      description: alphanumerical identifier for a resource
      type: string
      pattern: "^@[A-Z][0-9]{9}$"
      example: "@U000000001"
      readOnly: true
      minLength: 11
      maxLength: 11
    User:
      title: user
      description: local user
      type: object
      properties:
        Username:
          $ref: "#/components/schemas/Name"
        userId:
          $ref: "#/components/schemas/Identifier"
        UserPhoto:
          $ref: "#/components/schemas/Photo"
...